database deploy: throw error if vocabulary link expired

+++ database/load_csv.sh
@@ -3,5 +3,11 @@ apt-get install unzip wget -y
 
 cd tmp
 wget -O vocabulary.zip $VOC_URL
-unzip 'vocabulary.zip' -d vocabulary
-rm vocabulary.zip
+
+# Archive not empty
+if zipinfo vocabulary.zip > /dev/null; then
+	unzip 'vocabulary.zip' -d vocabulary
+	rm vocabulary.zip
+else 
+	raise error "Vocabulary link expired"
+fi
